+++ server/lib/deploy.js
@@ -1,28 +1,33 @@
 import { deploy as sourceDeploy } from 'auth0-source-control-extension-tools';
-import config from '../lib/config';
 
+import report from './reporter';
+import { getExcluded } from './storage';
+import config from './config';
 import { getChanges as getGitChanges } from './tfs-git';
 import { getChanges as getTfvcChanges } from './tfs-tfvc';
 
 export default (storage, id, repositoryId, branch, repository, sha, user, client) => {
   const getChanges = config('TFS_TYPE') === 'git' ? getGitChanges : getTfvcChanges;
 
-  const context = {
-    init: () => getChanges(repositoryId, sha)
-      .then(data => {
-        context.pages = data.pages;
-        context.rules = data.rules;
-        context.databases = data.databases;
-        context.clients = data.clients;
-        context.ruleConfigs = data.ruleConfigs;
-        context.resourceServers = data.resourceServers;
-      })
+  const repo = {
+    id,
+    sha,
+    user,
+    branch,
+    repository
   };
 
-  const slackTemplate = {
-    fallback: 'Visual Studio to Auth0 Deployment',
-    text: `Visual Studio (${config('TFS_TYPE')}) to Auth0 Deployment`
-  };
+  return getChanges(repositoryId, sha)
+    .then(assets =>
+      getExcluded(storage)
+        .then((exclude) => {
+          assets.exclude = exclude;
+          repo.assets = assets;
 
-  return sourceDeploy({ id, branch, repository, sha, user }, context, client, storage, config, slackTemplate);
+          return assets;
+        }))
+    .then(assets => sourceDeploy(assets, client, config))
+    .then(progress => report(storage, { repo, progress }))
+    .catch(err => report(storage, { repo, error: err.message })
+      .then(() => Promise.reject(err)));
 };